fixed unhandled exception

+++ backend/Models/LoanItem.cs
@@ -1,11 +1,11 @@
 public class LoanItem {
     //input parameters:
-    public int Id { get; set; }
-    public double Balance { get; set; }
-    public int Term { get; set; }
-    public double Rate { get; set; }
+    public long Id { get; set; }
+    public long Balance { get; set; }
+    public long Term { get; set; }
+    public long Rate { get; set; }
 
     //calculated parameters:
-    public double MonthlyPayment { get; set; } //When this is assigned wherever, call the Monthly payment function from LoanMath.cs
+    public long MonthlyPayment { get; set; } //When this is assigned wherever, call the Monthly payment function from LoanMath.cs
 
 }
